requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
braceexpand==0.1.7
chardet==4.0.0
click==8.0.3
colorama==0.4.4
discord-ext-bot==1.0.1
discord.py==1.7.3
dnspython==2.1.0
idna==3.3
import-expression==1.1.4
jishaku==2.3.2
multidict==5.2.0
pymongo==4.0.1
python-decouple==3.6
six==1.16.0
typing_extensions==4.0.1
yarl==1.7.2
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.8
